automation/requirements.txt
pandas>=2.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pytest>=8.0.0
httpx>=0.27.0
openpyxl>=3.1.2
pyxlsb>=1.0.10
xlrd>=2.0.1
fastapi>=0.109.0
uvicorn>=0.27.0
